feat(solana): puts BashMerkleTree within shared
