+++ packages/solana/shared/batch-merkle-tree/Cargo.toml
@@ -1,12 +1,8 @@
 [package]
-name = "snapshotter"
+name = "batch-merkle-tree"
 version = "0.1.0"
-description = "Created with Anchor"
+description = "A merkle tree that allows to calculate a root over a batch of leaves"
 edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "snapshotter"
 
 [features]
 default = []
@@ -17,5 +13,6 @@ no-log-ix-name = []
 idl-build = ["anchor-lang/idl-build"]
 
 [dependencies]
+alloy-sol-types = "0.7.0"
+alloy-primitives = "0.7.0"
 anchor-lang = "0.30.1"
-batch-merkle-tree = { path = "../../shared/batch-merkle-tree" }
